LES_G14/codigo/requirements.txt
asgiref==3.2.3
autopep8==1.5
Babel==2.8.0
cssselect==1.1.0
Django==3.0.7
django-environ==0.4.5
django-filter==2.2.0
django-formtools==2.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-phonenumber-field==4.0.0
django-selenium-login==2.0.0
django-tables2==2.3.1
djangorestframework==3.11.0
html5lib==1.0.1
jsonfield==3.1.0
lxml==4.5.1
mysqlclient==1.4.6
phonenumbers==8.11.5
Pillow==7.1.2
pycodestyle==2.5.0
PyPDF2==1.26.0
pyquery==1.4.1
pytz==2019.3
reportlab==3.5.42
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.4
